Ensure percentA() and percentB() don't truncate the floating-point division.

+++ jplag/src/main/java/jplag/AllMatches.java
@@ -106,13 +106,13 @@ public class AllMatches extends Matches implements Comparator<AllMatches> {
   	int divisor;
   	if(bcmatchesA != null) divisor = subA.size()-subA.files.length-bcmatchesA.tokensMatched();
     else divisor = subA.size()-subA.files.length;
-    return (divisor == 0 ? 0 : (tokensMatched()*100 / divisor));
+    return (divisor == 0 ? 0f : (tokensMatched()*100 / (float) divisor));
   }
   public final float percentB() {
     int divisor;
 	if(bcmatchesB != null) divisor = subB.size()-subB.files.length-bcmatchesB.tokensMatched();
 	else divisor = subB.size()-subB.files.length;
-    return (divisor == 0 ? 0 : (tokensMatched()*100 / divisor));
+    return (divisor == 0 ? 0f : (tokensMatched()*100 / (float) divisor));
   }
 
   public final float roundedPercentMaxAB() {
